update installation shell scirpt on linux

+++ linux.sh
@@ -1,21 +1,46 @@
 #!/usr/bin/env bash
 
-# required
-mv ~/.config/nvim{,.bak}
+ok=0
+option=""
 
-# optional but recommended
-mv ~/.local/share/nvim{,.bak}
-mv ~/.local/state/nvim{,.bak}
-mv ~/.cache/nvim{,.bak}
+if [ $# -gt 0 ]; then
+    option=$1
+fi
+    
+if [ "$option" = "install" ]; then
+    $ok=1
 
-# Clone the starter
-git clone https://github.com/LazyVim/starter ~/.config/nvim
+    # required
+    mv ~/.config/nvim{,.bak}
 
-# Remove the .git folder, so you can add it to your own repo later
-rm -rf ~/.config/nvim/.git ~/.config/nvim/.gitignore
+    # optional but recommended
+    mv ~/.local/share/nvim{,.bak}
+    mv ~/.local/state/nvim{,.bak}
+    mv ~/.cache/nvim{,.bak}
 
-# Copy my lazyvim settings
-cp -r ./lua ~/.config/nvim
+    # Clone the starter
+    git clone https://github.com/LazyVim/starter ~/.config/nvim
 
-#Start Neovim!
-nvim
+    # Remove the .git folder, so you can add it to your own repo later
+    rm -rf ~/.config/nvim/.git ~/.config/nvim/.gitignore
+
+    # Copy my lazyvim settings
+    cp -r ./lua ~/.config/nvim
+
+    #Start Neovim!
+    nvim
+elif [ "$option" = "config" ]; then
+    $ok=1
+
+    # Copy my lazyvim settings
+    cp -r ./lua ~/.config/nvim
+fi
+
+if [ $ok -eq 0 ]; then
+    echo "USAGE: $0 [install|config]"
+    echo "1) install: install LazyVim starter template & set my configuration"
+    echo "2) config: only set my configuration without installation of LazyVim starter template"
+    exit 1
+fi
+
+
